added more percision to strength of drink and other styling elements

+++ backend/db/migrations/20211206173338-create-drink.js
@@ -22,7 +22,7 @@ module.exports = {
       },
       strength: {
         allowNull: false,
-        type: Sequelize.INTEGER
+        type: Sequelize.DECIMAL(10, 1)
       },
       userId: {
         allowNull: false,
